fix actually - use pts for indexing frames not time stamps stuff works but kinda slow still

+++ repos/synapse/src/synapse/video_loader/loader.py
@@ -63,7 +63,7 @@ async def process_video(
                     shape=(
                         stream_metadata.output_video.total_frames,
                     ),  # Start with 0 timestamps
-                    dtype=ts.float64,
+                    dtype=ts.uint64,
                     path=args.output_path + "/timestamps",
                 ),
             )
